Use tox for travis runs
Travis has a matrix feature which can be used together with tox. That way,
there is only one place that defines the test environment.
Also fix the rst2html.py call - the parameter it needs is README.rst.
Also use containers for build env which is faster (sudo=False)
Also remove the py26 testrun. Python 2.6 is not supported anymore.