Update method signature in README.rst
The TestResourceManager implementation of `make()` actually requires
at least one argument, `dependency_resources`. The introductory
documentation in README.rst builds out an example, but the method
signature of `make()` doesn't include this required argument.

This commit updates the documentation to inlucde the required argument
to be consistent with the implementation. It also calls out the
example in doc/ by referencing it with a link.